# Optimize projectiles