--- a/lua/sim/DefaultProjectiles.lua
+++ b/lua/sim/DefaultProjectiles.lua
@@ -500,3 +500,4 @@ local DefaultProjectilesOpti = import("/lua/sim/DefaultProjectilesOpti.lua")
 
 EmitterProjectileOpti = DefaultProjectilesOpti.EmitterProjectile
 SingleBeamProjectileOpti = DefaultProjectilesOpti.SingleBeamProjectile
+MultiPolyTrailProjectileOpti = DefaultProjectilesOpti.MultiPolyTrailProjectile