--- a/lua/sim/defaultweapons.lua
+++ b/lua/sim/defaultweapons.lua
@@ -1333,3 +1333,6 @@ SCUDeathWeapon = Class(BareBonesWeapon) {
         myProjectile:PassDamageData(self:GetDamageTable())
     end,
 }
+
+-- Tactical missile threads --
+