--- a/projectiles/SIFThunthoArtilleryShell01/SIFThunthoArtilleryShell01_script.lua
+++ b/projectiles/SIFThunthoArtilleryShell01/SIFThunthoArtilleryShell01_script.lua
@@ -1,64 +1,75 @@
-#****************************************************************************
-#**
-#**  File     :  /data/projectiles/SIFThunthoArtilleryShell01/SIFThunthoArtilleryShell01_script.lua
-#**  Author(s):  Gordon Duclos, Aaron Lundquist
-#**
-#**  Summary  :  Thuntho Artillery Shell Projectile script, XSL0103
-#**
-#**  Copyright © 2007 Gas Powered Games, Inc.  All rights reserved.
-#****************************************************************************
+--****************************************************************************
+--**
+--**  File     :  /data/projectiles/SIFThunthoArtilleryShell01/SIFThunthoArtilleryShell01_script.lua
+--**  Author(s):  Gordon Duclos, Aaron Lundquist
+--**
+--**  Summary  :  Thuntho Artillery Shell Projectile script, XSL0103
+--**
+--**  Copyright ï¿½ 2007 Gas Powered Games, Inc.  All rights reserved.
+--****************************************************************************
 
-local EffectTemplate = import('/lua/EffectTemplates.lua')
-local SThunthoArtilleryShell = import('/lua/seraphimprojectiles.lua').SThunthoArtilleryShell
+local SThunderStormCannonProjectileSplitFx = import('/lua/EffectTemplates.lua').SThunderStormCannonProjectileSplitFx
+local SThunthoArtilleryShell = import('/lua/seraphimprojectiles.lua').SThunthoArtilleryShellOpti
 local RandomFloat = import('/lua/utilities.lua').GetRandomFloat
 local VizMarker = import('/lua/sim/VizMarker.lua').VizMarker
+
+-- globals as upvalues for performance 
+local Random = Random
+local CreateEmitterAtEntity = CreateEmitterAtEntity
+
+-- math functions as upvalues for performance
+local MathSin = _G.math.sin
+local MathCos = _G.math.cos 
+
+-- moho functions as upvalue for performance
+local EntityMethods = _G.moho.entity_methods
+local EntityDestroy = EntityMethods.Destroy
+
+local ProjectileMethods = _G.moho.projectile_methods
+local ProjectileGetVelocity = ProjectileMethods.GetVelocity
+local ProjectileSetVelocity = ProjectileMethods.SetVelocity
+local ProjectileCreateChildProjectile = ProjectileMethods.CreateChildProjectile
+
+-- attach for CTRL + SHIFT F replacement
 
 SIFThunthoArtilleryShell01 = Class(SThunthoArtilleryShell) {
                
     OnImpact = function(self, TargetType, TargetEntity) 
         
-        local FxFragEffect = EffectTemplate.SThunderStormCannonProjectileSplitFx 
-        local bp = self:GetBlueprint().Physics
+        local army = self.Army
+        local bp = self.Blueprint.Physics
+        local bpFragments = bp.Fragments
+        local bpFragmentId = bp.FragmentId
+        local damageData = self.DamageData
+        local FxFragEffect = SThunderStormCannonProjectileSplitFx
               
-        ### Split effects
+        -- split effects
         for k, v in FxFragEffect do
-            CreateEmitterAtEntity( self, self:GetArmy(), v )
-        end
-        
-        local vx, vy, vz = self:GetVelocity()
-        local velocity = 18
-    
-		# One initial projectile following same directional path as the original
-        #self:CreateChildProjectile(bp.FragmentId):SetVelocity(vx, vy, vz):SetVelocity(velocity):PassDamageData(self.DamageData)
-   		
-		# Create several other projectiles in a dispersal pattern
-        local numProjectiles = bp.Fragments
-        
-        local angle = (2 * math.pi) / numProjectiles
-        local angleInitial = RandomFloat( 0, angle )
-        
-        # Randomization of the spread
-        local angleVariation = angle * 0.8 # Adjusts angle variance spread
-        local spreadMul = 0.15 # Adjusts the width of the dispersal        
-        
-        --vy= -0.8
-        
+            CreateEmitterAtEntity( self, army, v )
+        end     
+
+        -- Randomization of the spread
+        local angle = (2 * 3.141592) / bpFragments
+        local angleInitial = Random() * angle
+        local angleVariation = angle * 0.8 -- Adjusts angle variance spread     
+
+        local vx, vy, vz = ProjectileGetVelocity(self)
+
         local xVec = 0
         local yVec = vy
         local zVec = 0
-     
 
-        # Launch projectiles at semi-random angles away from split location
-        for i = 0, numProjectiles - 1 do
-            xVec = vx + (math.sin(angleInitial + (i*angle) + RandomFloat(-angleVariation, angleVariation))) * spreadMul
-            zVec = vz + (math.cos(angleInitial + (i*angle) + RandomFloat(-angleVariation, angleVariation))) * spreadMul 
-            local proj = self:CreateChildProjectile(bp.FragmentId)
-            proj:SetVelocity(xVec,yVec,zVec)
-            proj:SetVelocity(velocity)
-            proj:PassDamageData(self.DamageData)                        
+        -- Launch projectiles at semi-random angles away from split location
+        for i = 1, bpFragments do
+            xVec = vx + (MathSin(angleInitial + (i*angle) + RandomFloat(-angleVariation, angleVariation))) * 0.15 -- spreadMul
+            zVec = vz + (MathCos(angleInitial + (i*angle) + RandomFloat(-angleVariation, angleVariation))) * 0.15 -- spreadMul 
+            local proj = ProjectileCreateChildProjectile(self, bpFragmentId)
+            ProjectileSetVelocity(proj, xVec, yVec, zVec)
+            ProjectileSetVelocity(proj, 18) -- velocity
+            proj.PassDamageData(proj, damageData)                        
         end
         
-        self:Destroy()
+        EntityDestroy(self)
     end
 }