… projectiles/SIFThunthoArtilleryShell02/SIFThunthoArtilleryShell02_script.lua
… projectiles/TIFMissileCruise04/TIFMissileCruise04_Script.lua
… projectiles/TIFMissileCruise04/TIFMissileCruise04_proj.bp
… units/UES0202/UES0202_script.lua