@Garanas 2021-09-12
Closing this PR, conclusion: creating a separate hierarchy next to the original will cause a lot more issues than it solves:
 - Mods that hook into the original hierarchy somehow fail silently.
 - Mods that hook into units and expect the original hierarchy fail.

It is better to just update the current hierarchy instead, taking into account that it will break mods and prevent that as much as we can.